requirements.txt
pillow==11.0